requirements.txt
python-telegram-bot==21.5
tronpy
aiohttp
python-dotenv
